[#525] feat: 동아리 생성 신청 formData의 applicantName을 세션에서 받아오도록 구현 및 API 연동
