+++ src/features/club-create/model/type.ts
@@ -1,0 +1,9 @@
+export interface ClubCreateFormData {
+  universityCode: string;
+  clubName: string;
+  clubCategory: string;
+  clubAffiliation: string;
+  logo: string;
+  instagram: string;
+  description: string;
+}
